+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Use [lightningcss](https://github.com/parcel-bundler/lightningcss) instead of css-minify, which is better maintained.
   - BREAKING: The `minify_css_level_*` Cfg options no longer apply and have been removed.
+- [onepass] Implement `Display` and `Error` for `Error` and `FriendlyError` structs.
 
 ## 0.12.0